chore: a bit of refactoring
